+++ src/View.java
@@ -5,7 +5,7 @@ import javax.swing.text.JTextComponent;
 //Interface for use of template method.
 public interface View {
 
-	public void update(ArrayList<Integer> numbers);
+	public void update(ArrayList<Double> numbers);
 	public void reset();
 	public JTextComponent show();
 	public String getName();
